% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Plots the spatial distribution of theta in a plane}
\usage{
gradient(theta, long = NULL, lat = NULL)
}
\arguments{
\item{theta}{a matrix with the theta values of all samples.}

\item{long}{a matrix or vector containing the longitude coordinates of the samples in the same order as theta}

\item{lat}{a matrix or vector containing the latitude coordinates of the samples in the same order as theta}
}
\value{
the plot of theta for every sample displayed with its coordinates
}
\description{
Plots the spatial distribution of theta in a plane
}
