% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutralitytest.R
\name{neutralitytest}
\alias{neutralitytest}
\title{Computes the difference between the theta of a subsample and the theta of another or the whole sample.}
\usage{
neutralitytest(vcf, s1, s2 = NULL, omega_type = "S", folded = FALSE)
}
\arguments{
\item{vcf}{A vcf file.}

\item{s1}{The name of the first sample to compare. A string object. It is necessary.}

\item{s2}{The name of the second sample to compare. A string object. NULL by default and the whole theta of the whole sample is compared.}

\item{omega_type}{The model to follow for omega ("S", "pi", "xi1", "H"). "S" is set as default.}

\item{folded}{A logical variable. If true the sfs is considered as folded (no knowledge over which is the alternative variant).}
}
\value{
The numerical difference between the thetas of both samples inserted.
}
\description{
Computes the difference between the theta of a subsample and the theta of another or the whole sample.
}
