% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfsmatrix.R
\name{rsfsmatrix}
\alias{rsfsmatrix}
\title{Calculate the rsfs matrix for every sample and every polymorphic site}
\usage{
rsfsmatrix(vcf)
}
\arguments{
\item{vcf}{A vcf file}
}
\value{
The rsfs matrix
}
\description{
Calculate the rsfs matrix for every sample and every polymorphic site
}
