% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{rtheta}
\alias{rtheta}
\title{Compute theta for subsamples of genetic sequences using the relative Site Frequency Spectrum (rSFS).}
\usage{
rtheta(vcf, omega_type = "S", folded = FALSE, subsample = "all")
}
\arguments{
\item{vcf}{A vcf file.}

\item{omega_type}{The model to follow for omega ("S", "pi", "xi1", "H"). "S" is set as default.}

\item{folded}{A logical variable. If true the rsfs is considered as folded (no knowledge over which is the alternative variant).}

\item{subsample}{The name of the subsample to compute theta for ("example"). "all" set as default.}
}
\value{
The numerical value of theta for a subsample or a vector with the theta values for each subsample.
}
\description{
Compute theta for subsamples of genetic sequences using the relative Site Frequency Spectrum (rSFS).
}
