% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfsmatrix.R
\name{sfsbarplot}
\alias{sfsbarplot}
\title{Plot the Site Frequency Spectrum into a barplot, representing the times an alternative allele frequency appears.}
\usage{
sfsbarplot(vcf)
}
\arguments{
\item{vcf}{A vcf file}
}
\value{
The barplot of SFS
}
\description{
Plot the Site Frequency Spectrum into a barplot, representing the times an alternative allele frequency appears.
}
