% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfsmatrix.R
\name{sfsmatrix}
\alias{sfsmatrix}
\title{Calculate the SFS matrix of frequencies or its contingency table}
\usage{
sfsmatrix(vcf, cont_tab = FALSE)
}
\arguments{
\item{vcf}{A vcf file}

\item{cont_tab}{A logical variable. If TRUE the function returns the contingency table of SFS as a matrix. If FALSE it returns the SFS matrix.}
}
\value{
The SFS matrix or its contingency table
}
\description{
Calculate the SFS matrix of frequencies or its contingency table
}
