% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{Compute theta for a complete sample of genetic sequences using the Site Frequency Spectrum (SFS).}
\usage{
theta(vcf, omega_type = "S", folded = FALSE)
}
\arguments{
\item{vcf}{A vcf file.}

\item{omega_type}{The model to follow for omega ("S", "pi", "xi1", "H"). "S" is set as default.}

\item{folded}{A logical variable. If true the sfs is considered as folded (no knowledge over which is the alternative variant).}
}
\value{
The numerical value of theta
}
\description{
Compute theta for a complete sample of genetic sequences using the Site Frequency Spectrum (SFS).
}
